// src/components/BetterAuth/SignUp/SignUp.interface.ts
import z from "zod";

export const signUpSchema = z
  .object({
    firstName: z.string().min(2, "Full name must be at least 2 characters"),
    lastName: z.string().min(2, "Full name must be at least 2 characters"),
    email: z.string().email("Invalid email address"),
    phone: z.string().min(10, "Phone number must be at least 10 digits"),
    password: z.string().min(6, "Password must be at least 6 characters"),
    confirmPassword: z.string(),
    photo: z.any(),
  })
  .refine((data) => data.password === data.confirmPassword, {
    message: "Passwords don't match",
    path: ["confirmPassword"],
  });

export type SignUpFormData = z.infer<typeof signUpSchema>;